Improving some little things

+++ becomer2.c
@@ -6,7 +6,7 @@
 /*   By: jleon-la <jleon-la@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/06 12:44:23 by jleon-la          #+#    #+#             */
-/*   Updated: 2024/02/06 17:52:21 by jleon-la         ###   ########.fr       */
+/*   Updated: 2024/02/06 18:50:34 by jleon-la         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	ft_address(unsigned long long int address, char *base)
 
 	i = 0;
 	if (address >= 16)
-		i += ft_address((unsigned long int)address / 16, base);
+		i += ft_address(address / 16, base);
 	i += write(1, &base[address % 16], 1);
 	return (i);
 }
